# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Segment Train Booking API
  version: 1.0.0
  summary: Segment-aware reserved-seat booking contract
  description: Initial documentation contract. Demo data is not official Sri Lanka Railways data.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/passenger/register:
    post:
      tags: [Passenger accounts]
      operationId: registerPassenger
      description: Creates an optional passenger account and starts an HTTP-only cookie session. Guest checkout remains available.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PassengerRegistration'}}}}
      responses:
        '201': {description: Passenger account created, content: {application/json: {schema: {$ref: '#/components/schemas/PassengerAccount'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
  /api/v1/passenger/login:
    post:
      tags: [Passenger accounts]
      operationId: loginPassenger
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PassengerLogin'}}}}
      responses:
        '200': {description: Signed-in passenger, content: {application/json: {schema: {$ref: '#/components/schemas/PassengerAccount'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
...
  /api/v1/passenger/me:
    get:
      tags: [Passenger accounts]
      operationId: getCurrentPassenger
      security: [{passengerCookie: []}]
      responses:
        '200': {description: Current passenger, content: {application/json: {schema: {$ref: '#/components/schemas/PassengerAccount'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/payments/sandbox:
    post:
      tags: [Bookings]
      operationId: checkoutSandboxPayment
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      description: Local-development payment adapter that atomically pays, confirms, issues a ticket, and queues notification delivery.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SandboxCheckoutRequest'}}}}
      responses:
        '201': {description: Paid booking and ticket, content: {application/json: {schema: {$ref: '#/components/schemas/CheckoutResult'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/payments/payhere:
    post:
      tags: [Payments]
      operationId: startPayHereCheckout
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      description: Creates or reuses a pending PayHere checkout. The returned fields form a server-signed POST to the hosted checkout page; no merchant secret is returned.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PayHereCheckoutRequest'}}}}
      responses:
        '201': {description: Signed hosted-checkout form, content: {application/json: {schema: {$ref: '#/components/schemas/PayHereCheckoutResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/payments/{paymentId}:
    get:
      tags: [Payments]
      operationId: getPayHerePayment
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/PaymentId'}]
      description: Returns the authoritative payment and booking state after the browser returns from hosted checkout.
      responses:
        '200': {description: Current payment state and ticket when paid, content: {application/json: {schema: {$ref: '#/components/schemas/PayHerePaymentStatus'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/webhooks/payhere:
    post:
      tags: [Payments]
      operationId: receivePayHereNotification
      description: PayHere server callback. The API verifies the provider signature, amount, currency and replay fingerprint before changing durable state.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: {$ref: '#/components/schemas/PayHereNotification'}
      responses:
        '200': {description: Notification processed or safely deduplicated}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/tickets/verify:
    post:
      tags: [Bookings]
      operationId: verifyTicket
      description: Verifies a ticket credential without exposing passenger personal data.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/VerifyTicketRequest'}}}}
      responses:
        '200': {description: Ticket verification result, content: {application/json: {schema: {$ref: '#/components/schemas/TicketVerification'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: BookingManagementToken, description: Short-lived signed token issued after contact verification}
    adminAuth: {type: http, scheme: bearer, bearerFormat: AdminAPIKey, description: Server-managed department administrator credential}
    passengerCookie: {type: apiKey, in: cookie, name: rail_passenger_session, description: Opaque revocable passenger session stored in an HTTP-only SameSite cookie}
  parameters:
...
    PaymentId: {name: paymentId, in: path, required: true, schema: {type: string, format: uuid}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
...
  responses:
...
    BadRequest: {description: Invalid input, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Unauthorized: {description: Authentication required or invalid, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Conflict: {description: Current state conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Unprocessable: {description: Domain rule violation, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    TooManyRequests: {description: Rate limit exceeded, headers: {Retry-After: {schema: {type: integer}}}, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    ServiceUnavailable: {description: Dependency unavailable, headers: {Retry-After: {schema: {type: integer}}}, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
...
    Seat:
      type: object
      required: [id, label, coachId, coachCode, coachClass, attributes]
      properties:
        id: {type: string, format: uuid}
        label: {type: string, example: 12A}
        coachId: {type: string, format: uuid}
        coachCode: {type: string, example: R1}
        coachClass: {type: string, example: FIRST}
        attributes: {type: array, items: {type: string}, example: [WINDOW]}
...
    Money:
      type: object
      required: [amountMinor, currency, currencyScale]
      properties: {amountMinor: {type: integer, format: int64, minimum: 0, example: 46000}, currency: {type: string, pattern: '^[A-Z]{3}$', example: LKR}, currencyScale: {type: integer, minimum: 0, maximum: 4, example: 2}}
...
    PassengerRegistration:
      type: object
      additionalProperties: false
      required: [fullName, email, password]
      properties:
        fullName: {type: string, minLength: 2, maxLength: 120}
        email: {type: string, format: email, maxLength: 254}
        phone: {type: string, pattern: '^\+[1-9]\d{7,14}$'}
        password: {type: string, minLength: 10, maxLength: 72, format: password}
    PassengerLogin:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, minLength: 1, maxLength: 72, format: password}
    PassengerAccount:
      type: object
      required: [id, fullName, email, createdAt]
      properties:
        id: {type: string, format: uuid}
        fullName: {type: string}
        email: {type: string, format: email}
        phone: {type: string}
        createdAt: {type: string, format: date-time}
...
    SandboxCheckoutRequest:
      type: object
      additionalProperties: false
      required: [bookingId, bookingToken]
      properties:
        bookingId: {type: string, format: uuid}
        bookingToken: {type: string}
    PayHereCheckoutRequest:
      type: object
      additionalProperties: false
      required: [bookingId, bookingToken, billingAddress, city]
      properties:
        bookingId: {type: string, format: uuid}
        bookingToken: {type: string}
        billingAddress: {type: string, minLength: 3, maxLength: 200}
        city: {type: string, minLength: 2, maxLength: 80}
    PayHereCheckoutResponse:
      type: object
      required: [paymentId, bookingId, status, expiresAt, actionUrl, fields]
      properties:
        paymentId: {type: string, format: uuid}
        bookingId: {type: string, format: uuid}
        status: {type: string, const: PENDING}
        expiresAt: {type: string, format: date-time}
        actionUrl: {type: string, format: uri, description: PayHere hosted Sandbox or production checkout endpoint selected by server configuration}
        fields: {type: object, additionalProperties: {type: string}, description: Signed hidden form fields to POST to actionUrl}
    PayHerePaymentStatus:
      type: object
      required: [paymentId, bookingId, status, providerReference, amountMinor, currency, booking]
      properties:
        paymentId: {type: string, format: uuid}
        bookingId: {type: string, format: uuid}
        status: {type: string, enum: [PENDING, PAID, FAILED, DISPUTED]}
        providerReference: {type: string}
        amountMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        paidAt: {type: string, format: date-time}
        booking: {$ref: '#/components/schemas/Booking'}
        ticket: {type: object, required: [id, verificationCode, status], properties: {id: {type: string, format: uuid}, verificationCode: {type: string}, status: {type: string, enum: [ACTIVE, CANCELLED, USED]}}}
    PayHereNotification:
      type: object
      additionalProperties: true
      required: [merchant_id, order_id, payment_id, payhere_amount, payhere_currency, status_code, md5sig]
      properties:
        merchant_id: {type: string}
        order_id: {type: string, format: uuid}
        payment_id: {type: string}
        payhere_amount: {type: string, pattern: '^\\d+\\.\\d{2}$'}
        payhere_currency: {type: string, pattern: '^[A-Z]{3}$'}
        status_code: {type: string, enum: ['2', '0', '-1', '-2', '-3']}
        md5sig: {type: string, pattern: '^[A-F0-9]{32}$'}
    CheckoutResult:
      type: object
      required: [payment, ticket, booking]
      properties:
        payment: {type: object, additionalProperties: true}
        ticket: {type: object, required: [id, verificationCode, status], properties: {id: {type: string, format: uuid}, verificationCode: {type: string}, status: {type: string, enum: [ACTIVE, CANCELLED, USED]}}}
        booking: {$ref: '#/components/schemas/Booking'}
    VerifyTicketRequest:
      type: object
      additionalProperties: false
      required: [verificationCode]
      properties: {verificationCode: {type: string, minLength: 32, maxLength: 128}}
    TicketVerification:
      type: object
      required: [valid, ticketId, ticketStatus, bookingReference, bookingStatus, trainRunId, coachCode, seatLabel, originStationId, destinationStationId, verifiedAt]
      properties:
        valid: {type: boolean}
        ticketId: {type: string, format: uuid}
        ticketStatus: {type: string, enum: [ACTIVE, CANCELLED, USED]}
        bookingReference: {type: string}
        bookingStatus: {type: string}
        trainRunId: {type: string, format: uuid}
        coachCode: {type: string}
        seatLabel: {type: string}
        originStationId: {type: string, format: uuid}
        destinationStationId: {type: string, format: uuid}
        verifiedAt: {type: string, format: date-time}
    Booking:
      type: object
      required: [id, reference, status, trainRunId, seat, originStationId, destinationStationId, fare, createdAt]
      properties:
        id: {type: string, format: uuid}
        reference: {type: string, example: BK-7J4M9Q2X}
        status: {type: string, enum: [PENDING, HELD, CONFIRMED, CANCELLED, EXPIRED, COMPLETED]}
        trainRunId: {type: string, format: uuid}
        seat: {$ref: '#/components/schemas/Seat'}
        originStationId: {type: string, format: uuid}
        destinationStationId: {type: string, format: uuid}
        fare: {$ref: '#/components/schemas/Money'}
        createdAt: {type: string, format: date-time}
        confirmedAt: {type: [string, 'null'], format: date-time}
        cancelledAt: {type: [string, 'null'], format: date-time}
        managementToken: {type: string, description: Short-lived opaque token for managing this booking}
        refund: {$ref: '#/components/schemas/RefundSummary'}
...
    RefundSummary:
      type: object
      required: [id, status, amountMinor, currency]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [PENDING, SUCCEEDED, FAILED]}
        amountMinor: {type: integer, format: int64}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
...
    FieldError:
      type: object
      required: [field, message]
      properties: {field: {type: string}, message: {type: string}}
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: {type: string, example: VALIDATION_ERROR}
        message: {type: string, example: The request contains invalid fields.}
        details: {oneOf: [{type: array, items: {$ref: '#/components/schemas/FieldError'}}, {type: object, additionalProperties: true}]}
        requestId: {type: string, format: uuid}